+++ Docs/API/html/search/all_b.js
@@ -1,7 +1,7 @@
 var searchData=
 [
   ['reallocfn',['reallocFn',['../struct_jinx_1_1_global_params.html#ae135b2616b7c2df0f5ec8e12b3f1c9e3',1,'Jinx::GlobalParams::reallocFn()'],['../_jinx_8h.html#adbe71f891fedec537a90904bb32bb111',1,'Jinx::ReallocFn()']]],
-  ['registerfunction',['RegisterFunction',['../class_jinx_1_1_i_library.html#a7fe80895d328feaef272bda6cec3b5d9',1,'Jinx::ILibrary']]],
-  ['registerproperty',['RegisterProperty',['../class_jinx_1_1_i_library.html#a968f0464cf24d791920563af00703179',1,'Jinx::ILibrary']]],
+  ['registerfunction',['RegisterFunction',['../class_jinx_1_1_i_library.html#a1f9d85a1be41a6ebc482b698f038f76e',1,'Jinx::ILibrary']]],
+  ['registerproperty',['RegisterProperty',['../class_jinx_1_1_i_library.html#adf209831e54668f48fc9f241cb545825',1,'Jinx::ILibrary']]],
   ['runtimeptr',['RuntimePtr',['../_jinx_8h.html#a6bc9758497dff930dae7356ec5a43a42',1,'Jinx']]]
 ];