+++ Source/JxLibrary.cpp
@@ -128,8 +128,7 @@ FunctionSignature Library::CreateFunctionSignature(bool publicScope, bool return
 	}
 
 	// Create functions signature
-	VisibilityType scope = publicScope ? VisibilityType::Public : VisibilityType::Private;
-	FunctionSignature functionSignature(scope, returnValue, GetName(), parts);
+	FunctionSignature functionSignature(publicScope ? VisibilityType::Public : VisibilityType::Private, returnValue, GetName(), parts);
 	return functionSignature;
 }
 
@@ -160,7 +159,7 @@ bool Library::PropertyNameExists(const String & name) const
 	return m_propertyNameTable.find(name) == m_propertyNameTable.end() ? false : true;
 }
 
-bool Library::RegisterFunction(bool publicScope, bool returnValue, std::initializer_list<String> name, FunctionCallback function)
+bool Library::RegisterFunction(Visibility visibility, ReturnValue returnValue, std::initializer_list<String> name, FunctionCallback function)
 {
 	if (name.size() < 1)
 	{
@@ -174,7 +173,7 @@ bool Library::RegisterFunction(bool publicScope, bool returnValue, std::initiali
 	}
 
 	// Calculate the function signature
-	FunctionSignature functionSignature = CreateFunctionSignature(publicScope, returnValue, name);
+	FunctionSignature functionSignature = CreateFunctionSignature(visibility == Visibility::Public ? true : false, returnValue == ReturnValue::Required ? true : false, name);
 	if (!functionSignature.IsValid())
 		return false;
 
@@ -193,12 +192,12 @@ bool Library::RegisterFunction(bool publicScope, bool returnValue, std::initiali
 	return true;
 }
 
-bool Library::RegisterProperty(bool readOnly, bool publicScope, const String & name, const Variant & value)
+bool Library::RegisterProperty(Visibility visibility, Access access, const String & name, const Variant & value)
 {
 	std::lock_guard<Mutex> lock(m_propertyMutex);
 
 	// Register the property name with the library
-	PropertyName prop(readOnly, publicScope ? VisibilityType::Public : VisibilityType::Private, GetName(), name);
+	PropertyName prop(visibility == Visibility::Public ? VisibilityType::Public : VisibilityType::Private, access == Access::ReadOnly ? true : false, GetName(), name);
 	if (!RegisterPropertyNameInternal(prop, false))
 		return false;