+++ Docs/Changelog.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 ## [0.7.0] - 2017-07-07
 
 ### Changed
-- Added convenience methods to support alternate string encodings more seamlessly
+- Added better support for alternative string character encodings / types
 - Changed function registration with more explicit parameters
+- Changed required ordering of readonly parameter for properties
 
 ## [0.6.0] - 2017-06-02